added arming, adapter drivers for my use

+++ Drivers/stm32-ssd1306/ssd1306.h
@@ -120,31 +120,31 @@ typedef enum {
 
 // Struct to store transformations
 typedef struct {
-    uint16_t CurrentX;
-    uint16_t CurrentY;
+    int16_t CurrentX;
+    int16_t CurrentY;
     uint8_t Initialized;
     uint8_t DisplayOn;
 } SSD1306_t;
 
 typedef struct {
-    uint8_t x;
-    uint8_t y;
+    int16_t x;
+    int16_t y;
 } SSD1306_VERTEX;
 
 // Procedure definitions
 void ssd1306_Init(void);
 void ssd1306_Fill(SSD1306_COLOR color);
 void ssd1306_UpdateScreen(void);
-void ssd1306_DrawPixel(uint8_t x, uint8_t y, SSD1306_COLOR color);
+void ssd1306_DrawPixel(int16_t x, int16_t y, SSD1306_COLOR color);
 char ssd1306_WriteChar(char ch, FontDef Font, SSD1306_COLOR color);
 char ssd1306_WriteString(char* str, FontDef Font, SSD1306_COLOR color);
-void ssd1306_SetCursor(uint8_t x, uint8_t y);
-void ssd1306_Line(uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2, SSD1306_COLOR color);
-void ssd1306_DrawArc(uint8_t x, uint8_t y, uint8_t radius, uint16_t start_angle, uint16_t sweep, SSD1306_COLOR color);
-void ssd1306_DrawCircle(uint8_t par_x, uint8_t par_y, uint8_t par_r, SSD1306_COLOR color);
+void ssd1306_SetCursor(int16_t x, int16_t y);
+void ssd1306_Line(int16_t x1, int16_t y1, int16_t x2, int16_t y2, SSD1306_COLOR color);
+void ssd1306_DrawArc(int16_t x, int16_t y, uint8_t radius, uint16_t start_angle, uint16_t sweep, SSD1306_COLOR color);
+void ssd1306_DrawCircle(int16_t par_x, int16_t par_y, uint8_t par_r, SSD1306_COLOR color);
 void ssd1306_Polyline(const SSD1306_VERTEX *par_vertex, uint16_t par_size, SSD1306_COLOR color);
-void ssd1306_DrawRectangle(uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2, SSD1306_COLOR color);
-void ssd1306_DrawBitmap(uint8_t x, uint8_t y, const unsigned char* bitmap, uint8_t w, uint8_t h, SSD1306_COLOR color);
+void ssd1306_DrawRectangle(int16_t x1, int16_t y1, int16_t x2, int16_t y2, SSD1306_COLOR color);
+void ssd1306_DrawBitmap(int16_t x, int16_t y, const unsigned char* bitmap, uint8_t w, uint8_t h, SSD1306_COLOR color);
 /**
  * @brief Sets the contrast of the display.
  * @param[in] value contrast to set.
